refactor: move the upgrade component to `Collection Filters` block
