+++ plugins/woocommerce-blocks/assets/js/blocks/collection-filters/edit.tsx
@@ -12,7 +12,7 @@ import type { AttributeSetting } from '@woocommerce/types';
  */
 import { EditProps, FilterType } from './types';
 import { getAllowedBlocks } from './utils';
-import Inspector from './components/inspector';
+import Downgrade from './components/downgrade';
 import Warning from './components/warning';
 import './editor.scss';
 
@@ -66,7 +66,7 @@ const Edit = ( props: EditProps ) => {
 	return (
 		<nav { ...blockProps }>
 			{ ! isNested && <Warning /> }
-			<Inspector clientId={ props.clientId } />
+			<Downgrade clientId={ props.clientId } />
 			<InnerBlocks
 				template={ template }
 				allowedBlocks={ allowedBlocks }
