+++ plugins/woocommerce-blocks/assets/js/blocks/filter-wrapper/types.ts
@@ -1,4 +1,11 @@
-export interface Attributes {
+/**
+ * External dependencies
+ */
+import { type BlockEditProps } from '@wordpress/blocks';
+
+export type Attributes = {
 	heading: string;
 	filterType: string;
-}
+};
+
+export type EditProps = BlockEditProps< Attributes >;
